Colunas de Autor, Created_At e Status de todos os CRUDS

+++ handlers/handlersFeatures.go
@@ -12,32 +12,32 @@ import (
 )
 
 func CreateFeatureHandler(w http.ResponseWriter, r *http.Request) {
-	sec.IsAuthenticated(w, r)
 	log.Println("Create Feature")
-	if r.Method == "POST" {
+	if r.Method == "POST" && sec.IsAuthenticated(w, r) {
 		name := r.FormValue("Name")
 		code := r.FormValue("Code")
-		sqlStatement := "INSERT INTO features(name, code) VALUES ($1, $2) RETURNING id"
+		description := r.FormValue("Description")
+		sqlStatement := "INSERT INTO features(name, code, description) VALUES ($1, $2, $3) RETURNING id"
 		id := 0
-		err := Db.QueryRow(sqlStatement, name, code).Scan(&id)
-		sec.CheckInternalServerError(err, w)
+		err := Db.QueryRow(sqlStatement, name, code, description).Scan(&id)
 		if err != nil {
 			panic(err.Error())
 		}
-		sec.CheckInternalServerError(err, w)
-		log.Println("INSERT: Id: " + strconv.Itoa(id) + " | Name: " + name + " | Code: " + code)
+		log.Println("INSERT: Id: " + strconv.Itoa(id) + " | Name: " + name + " | Code: " + code + " | Description: " + description)
+		http.Redirect(w, r, route.FeaturesRoute, 301)
+	} else {
+		http.Redirect(w, r, "/logout", 301)
 	}
-	http.Redirect(w, r, route.FeaturesRoute, 301)
 }
 
 func UpdateFeatureHandler(w http.ResponseWriter, r *http.Request) {
-	sec.IsAuthenticated(w, r)
 	log.Println("Update Feature")
-	if r.Method == "POST" {
+	if r.Method == "POST" && sec.IsAuthenticated(w, r) {
 		id := r.FormValue("Id")
 		name := r.FormValue("Name")
 		code := r.FormValue("Code")
-		sqlStatement := "UPDATE features SET name=$1, code=$2 WHERE id=$3"
+		description := r.FormValue("Description")
+		sqlStatement := "UPDATE features SET name=$1, code=$2, description=$3 WHERE id=$4"
 		updtForm, err := Db.Prepare(sqlStatement)
 		sec.CheckInternalServerError(err, w)
 		if err != nil {
@@ -45,15 +45,16 @@ func UpdateFeatureHandler(w http.ResponseWriter, r *http.Request) {
 		}
 		sec.CheckInternalServerError(err, w)
 		updtForm.Exec(name, code, id)
-		log.Println("UPDATE: Id: " + id + " | Name: " + name + " | Code: " + code)
+		log.Println("UPDATE: Id: " + id + " | Name: " + name + " | Code: " + code + " | Description: " + description)
+		http.Redirect(w, r, route.FeaturesRoute, 301)
+	} else {
+		http.Redirect(w, r, "/logout", 301)
 	}
-	http.Redirect(w, r, route.FeaturesRoute, 301)
 }
 
 func DeleteFeatureHandler(w http.ResponseWriter, r *http.Request) {
-	sec.IsAuthenticated(w, r)
 	log.Println("Delete Feature")
-	if r.Method == "POST" {
+	if r.Method == "POST" && sec.IsAuthenticated(w, r) {
 		id := r.FormValue("Id")
 		sqlStatement := "DELETE FROM features WHERE id=$1"
 		deleteForm, err := Db.Prepare(sqlStatement)
@@ -63,8 +64,10 @@ func DeleteFeatureHandler(w http.ResponseWriter, r *http.Request) {
 		deleteForm.Exec(id)
 		sec.CheckInternalServerError(err, w)
 		log.Println("DELETE: Id: " + id)
+		http.Redirect(w, r, route.FeaturesRoute, 301)
+	} else {
+		http.Redirect(w, r, "/logout", 301)
 	}
-	http.Redirect(w, r, route.FeaturesRoute, 301)
 }
 
 func DeleteFeaturesByRoleHandler(roleId string) {
@@ -90,28 +93,55 @@ func DeleteFeaturesHandler(diffDB []mdl.Feature) {
 
 func ListFeaturesHandler(w http.ResponseWriter, r *http.Request) {
 	log.Println("List Features")
-	sec.IsAuthenticated(w, r)
-	rows, err := Db.Query("SELECT id, name, code FROM features order by id asc")
-	sec.CheckInternalServerError(err, w)
-	var features []mdl.Feature
-	var feature mdl.Feature
-	var i = 1
-	for rows.Next() {
-		err = rows.Scan(&feature.Id, &feature.Name, &feature.Code)
-		sec.CheckInternalServerError(err, w)
-		feature.Order = i
-		i++
-		features = append(features, feature)
+	if sec.IsAuthenticated(w, r) {
+		sql := "SELECT " +
+			" a.id, " +
+			" a.name, " +
+			" a.code, " +
+			" coalesce(a.description,'') as dsc, " +
+			" a.author_id, " +
+			" b.name, " +
+			" to_char(a.created_at,'DD/MM/YYYY HH24:MI:SS'), " +
+			" coalesce(c.name,'') as cstatus, " +
+			" a.status_id, " +
+			" a.id_versao_origem " +
+			" FROM features a LEFT JOIN users b " +
+			" ON a.author_id = b.id " +
+			" LEFT JOIN status c ON a.status_id = c.id " +
+			" order by a.id asc"
+		log.Println(sql)
+		rows, _ := Db.Query(sql)
+		var features []mdl.Feature
+		var feature mdl.Feature
+		var i = 1
+		for rows.Next() {
+			rows.Scan(
+				&feature.Id,
+				&feature.Name,
+				&feature.Code,
+				&feature.Description,
+				&feature.AuthorId,
+				&feature.AuthorName,
+				&feature.C_CreatedAt,
+				&feature.CStatus,
+				&feature.StatusId,
+				&feature.IdVersaoOrigem)
+			log.Println(feature.AuthorName)
+			feature.Order = i
+			i++
+			features = append(features, feature)
+		}
+		var page mdl.PageFeatures
+		page.Features = features
+		page.AppName = mdl.AppName
+		page.Title = "Funcionalidades"
+		page.LoggedUser = BuildLoggedUser(GetUserInCookie(w, r))
+		var tmpl = template.Must(template.ParseGlob("tiles/features/*"))
+		tmpl.ParseGlob("tiles/*")
+		tmpl.ExecuteTemplate(w, "Main-Features", page)
+	} else {
+		http.Redirect(w, r, "/logout", 301)
 	}
-	var page mdl.PageFeatures
-	page.Features = features
-	page.AppName = mdl.AppName
-	page.Title = "Funcionalidades"
-	page.LoggedUser = BuildLoggedUser(GetUserInCookie(w, r))
-	var tmpl = template.Must(template.ParseGlob("tiles/features/*"))
-	tmpl.ParseGlob("tiles/*")
-	tmpl.ExecuteTemplate(w, "Main-Features", page)
-	sec.CheckInternalServerError(err, w)
 }
 
 // AJAX
